+++ mea/src/once/once_cell/mod.rs
@@ -86,12 +86,12 @@ impl<T> OnceCell<T> {
     }
 
     /// Returns whether the internal value is set.
-    fn initialized(&self) -> bool {
+    pub(crate) fn initialized(&self) -> bool {
         self.value_set.load(Ordering::Acquire)
     }
 
     /// Returns whether the internal value is set.
-    fn initialized_mut(&mut self) -> bool {
+    pub(crate) fn initialized_mut(&mut self) -> bool {
         *self.value_set.get_mut()
     }